refactor: お気にいりのゲーム取得中にstatus===loadingを取得できるようにする
